Replaced string module - Replaced usage of the string module.
