+++ bin/general-template/newstudytemplate/dynamic-query
@@ -1,6 +1,6 @@
 #! /usr/bin/env node
 
-// Copyright 2020-2021, SimTK DataShare Team
+// Copyright 2020-2022, SimTK DataShare Team
 //
 // This file is part of SimTK DataShare. Initial development
 // was funded under NIH grants R01GM107340 and U54EB020405
@@ -15,10 +15,6 @@
 var fs = require('fs');
 var jsdom = require('/usr/lib/nodejs/npm/node_modules/jsdom');
 const {JSDOM} = jsdom;
-var $S = require('/usr/lib/nodejs/npm/node_modules/string');
-
-// Let string library extend String JS object prototypes.
-$S.extendPrototype();
 
 // LOAD THE STUDY TEMPLATE AND THE METADATA HTML TREE.
 var template = process.argv.length == 3 ?
