+++ lib/general-template/general.rules.js
@@ -1,5 +1,5 @@
 /**
- * Copyright 2020-2021, SimTK DataShare Team
+ * Copyright 2020-2022, SimTK DataShare Team
  *
  * This file is part of SimTK DataShare. Initial development
  * was funded under NIH grants R01GM107340 and U54EB020405
@@ -7,15 +7,6 @@
  * W81XWH-15-1-0232R01. Continued maintenance and enhancement
  * are funded by NIH grant R01GM124443.
  */
-
-simtkHumanizeAttrName = function(attrName) {
-	var fieldName   = attrName.toString();
-	fieldName = fieldName.replace(/_/g, " ");
-	fieldName = fieldName.replace(/[^0-9a-z ]/gi, '').trim();
-	fieldName = fieldName.charAt(0).toUpperCase() + fieldName.slice(1);
-
-	return fieldName;
-}
 
 // For tracking number of type value mismatches.
 cntTypeValueMismatches = 0;
@@ -67,6 +58,7 @@ var GeneralTemplate = function() {
 		var units   = {};
 		var add   = { 
 			detailed : { filter: function( field, context ) {
+				require('/usr/local/mobilizeds/lib/general-template/common_utils.js');
 				var fieldname   = simtkHumanizeAttrName(field.attr('name'));
 				var id          = rules.id(( context.concat([ fieldname ])).join( ' > ' ));
 				var optgroup    = context[ 0 ];
