+++ lib/general-template/rules.js
@@ -1,5 +1,5 @@
 /**
- * Copyright 2020-2021, SimTK DataShare Team
+ * Copyright 2020-2022, SimTK DataShare Team
  *
  * This file is part of SimTK DataShare. Initial development
  * was funded under NIH grants R01GM107340 and U54EB020405
@@ -48,7 +48,11 @@ module.exports = function( $ ) {
 	/*
 		@brief Returns an array of human-readable parent directories of the given file
 	*/
-		var context = this.path( metadatafile ).map(( c ) => { return c.humanize().toString(); });
+		var context = this.path( metadatafile ).map(( c ) => {
+			require('/usr/local/mobilizeds/lib/general-template/common_utils.js');
+			var humanized = simtkHumanize(c);
+			return humanized;
+		});
 		if( context.length > 1 && context[ 0 ].match( /subject/i )) { context = context.slice( 1 ); }
 		return context;
 	};
